删除远端服务器上好友的关系

+++ src/cn/ucai/superwechat/activity/MainActivity.java
@@ -580,11 +580,35 @@ public class MainActivity extends BaseActivity implements EMEventListener {
 		public void onContactDeleted(final List<String> usernameList) {
 			// 被删除
 			Map<String, User> localUsers = ((DemoHXSDKHelper)HXSDKHelper.getInstance()).getContactList();
-			for (String username : usernameList) {
-				localUsers.remove(username);
-				userDao.deleteContact(username);
-				inviteMessgeDao.deleteMessage(username);
+			String currentName = SuperWeChatApplication.getInstance().getUserName();
+			for (final String username : usernameList) {
+					final OkHttpUtils2<Result> utils = new OkHttpUtils2<Result>();
+					utils.setRequestUrl(I.REQUEST_DELETE_CONTACT)
+							.addParam(I.Contact.USER_NAME,currentName)
+							.addParam(I.Contact.CU_NAME,username)
+							.targetClass(Result.class)
+							.execute(new OkHttpUtils2.OnCompleteListener<Result>() {
+								@Override
+								public void onSuccess(Result result) {
+									if (result.isRetMsg()) {
+										((DemoHXSDKHelper)HXSDKHelper.getInstance()).getContactList().remove(username);
+										UserAvatar u = SuperWeChatApplication.getInstance().getUserMap().get(username);
+										SuperWeChatApplication.getInstance().getUserList().remove(u);
+										SuperWeChatApplication.getInstance().getUserMap().remove(username);
+										userDao.deleteContact(username);
+										inviteMessgeDao.deleteMessage(username);
+										sendStickyBroadcast(new Intent("update_contact_list"));
+									}
+								}
+
+								@Override
+								public void onError(String error) {
+									Log.e(TAG,"error = "+error);
+								}
+							});
+
 			}
+
 			runOnUiThread(new Runnable() {
 				public void run() {
 					// 如果正在与此用户的聊天页面
